feat: checks now no longer need to implement the common check code
- most checks use the same code for performing the check; this has been pulled up to chk3's check class
- check now has a new member function to perform a lookup on one ASTContext
- fio is still weird

Issues: #7

+++ src/cg3-common/execute_process.cxx
@@ -35,7 +35,7 @@ cg3::execute_process(const std::vector<std::string_view>& args,
     std::uint8_t buf[io_buffer_size];
     while (input.read(reinterpret_cast<char*>(buf), std::size(buf))) {
         auto red = input.gcount();
-        proc.write(static_cast<std::uint8_t*>(buf), red);
+        proc.write(static_cast<std::uint8_t*>(buf), static_cast<std::size_t>(red));
     }
     proc.close(reproc::stream::in);
